% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrustModel.R
\name{run_sim}
\alias{run_sim}
\title{Run the simulation}
\usage{
run_sim(num_trans)
}
\description{
Simulation the trust model for the number of transactions
}
\keyword{model}
\keyword{run}
\keyword{simulation}
\keyword{trust}
